LibState -> Lib
- Moved LibState-related functions to Lib module
- Lib type stores current position as well as Pos hashmaps

Converting [Point] to Pos all the time is a bit annoying, but it's inevitable because moves in Pos are unordered. Maybe I should use a different Pos implementation with ordered moves and an oredr/mirror-agnostic hash function?

I don't quite like the naming of Move, Pos and Point - it does not make it clear what each of those are. What about Stone, MoveSet (or MoveSeq) and Move?