List -> Ordered Set for MoveSeq
Since random access (checking if a Move exists in a MoveSeq, etc.) actually *is* important for MoveSeq, I thought that using something other than List should result in a speedup.

Ordered Set seemed like a perfect fit: Moves in a MoveSeq are always unique, existence checks are O(log n), and move order is preserved.

Actually implementing it was a PITA for several reasons:
- it does not have an instance of Container class from Universum, and some of the default classes require it to be wrapped in Biased newtype (which I haven't figured out how to do properly, because Data.Set.Ordered does not export L and R types)
- if I want MoveSeq.stoneAt to be faster than O(n), I have to add new moves to the right, which makes it harder to treat MoveSeq as a list
- I don't know if indexing is faster than O(n)

In the end, my implementation has a bunch of conversions to / from a list, which makes it quite slow, if not slower than the original version.