Universum -> Relude
Even though Universum is nice to use, it pulls in a lot of unnecessary dependencies; In my case, Monomer depends on lens, while Universum depends on microlens, so I have to depend on both.

Relude is pretty good and has minimal dependencies, but it misses some nice things such as a safeHead function ([a] -> Maybe a) or mono-traversable re-export. I guess I'd have to write a custom Prelude of my own at some point.